Refactor fingerprinting into a Fingerprint module
Fingerprint::Sha1 and Fingerprint::MD5 instead of the
less descriptive Default and Legacy.

Updates README
