+++ README.md
@@ -409,12 +409,12 @@ Then get the `notifier.js` from `errbit/public/javascript/notifier.js` and add t
 Using custom fingerprinting methods
 -----------------------------------
 
-Errbit now allows you to easily use your own Fingerprint Strategy if that's what you'd like to do. If you are upgrading from a very old version of errbit, you can use the `LegacyFingerprint` to provide yourself
+Errbit now allows you to easily use your own Fingerprint Strategy if that's what you'd like to do. If you are upgrading from a very old version of errbit, you can use the `Fingerprint::MD5` to provide yourself
 with compatibility. The fingerprint strategy can be changed by adding an initializer to errbit:
 
 ```ruby
 # config/fingerprint.rb
-ErrorReport.fingerprint_strategy = LegacyFingerprint
+ErrorReport.fingerprint_strategy = Fingerprint::MD5
 ```
 
 The easiest way to add custom fingerprint methods is to simply subclass `Fingerprint`
